requirements.txt
streamlit
roboflow
pillow
google-generativeai
